backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
uvloop==0.21.0
httptools==0.6.4
python-socketio==5.11.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
stripe==11.4.1
httpx==0.28.1
python-multipart==0.0.19
cryptography==44.0.0
celery[redis]==5.4.0
boto3==1.36.3
Pillow==11.1.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
aiosqlite==0.20.0
